requirements.txt
streamlit
ultralytics
opencv-python-headless==4.8.0.76
Pillow
numpy==1.24.4
plotly